Profile emissions and upstream emissions

These functions wrap the output of the corresponding function in tiltIndicator.

Usage

profile_emissions(
  companies,
  co2,
  europages_companies,
  ecoinvent_activities,
  ecoinvent_europages,
  isic,
  isic_tilt = lifecycle::deprecated(),
  low_threshold = 1/3,
  high_threshold = 2/3
)

Arguments

companies, co2

A dataframe like the dataset with a matching name in tiltToyData (see Reference).

europages_companies

Dataframe. Companies from europages.

ecoinvent_activities

Dataframe. Activities from ecoinvent.

ecoinvent_europages

Dataframe. Mapper between europages and ecoinvent.

isic

Dataframe. ISIC data.

isic_tilt

low_threshold

A numeric value to segment low and medium emission profile products.

high_threshold

A numeric value to segment medium and high emission profile products.

Value

A data frame with the column companies_id, and the nested columnsproduct and company holding the outputs at product and company level. Unnesting product yields a data frame with at least columns companies_id, grouped_by, risk_category. Unnesting company yields a data frame with at least columns companies_id, grouped_by, risk_category, value.

See also

Examples

library(tiltToyData)
library(readr, warn.conflicts = FALSE)

options(readr.show_col_types = FALSE)

companies <- read_csv(toy_emissions_profile_any_companies())
products <- read_csv(toy_emissions_profile_products())
#> Warning: `toy_emissions_profile_products()` was deprecated in tiltToyData 0.0.0.9009.
#> ℹ Please use `toy_emissions_profile_products_ecoinvent()` instead.
europages_companies <- read_csv(toy_europages_companies())
ecoinvent_activities <- read_csv(toy_ecoinvent_activities())
ecoinvent_europages <- read_csv(toy_ecoinvent_europages())
isic_name <- read_csv(toy_isic_name())

result <- profile_emissions(
  companies,
  products,
  europages_companies = europages_companies,
  ecoinvent_activities = ecoinvent_activities,
  ecoinvent_europages = ecoinvent_europages,
  isic = isic_name
)
#> Warning: Splitting `companies` into 4 chunks.

result |> unnest_product()
#> # A tibble: 72 × 25
#>    companies_id       company_name country emission_profile benchmark ep_product
#>    <chr>              <chr>        <chr>   <chr>            <lgl>     <chr>     
#>  1 asteria_megalotom… asteria_meg… austria NA               NA        NA        
#>  2 skarn_gallinule    skarn_galli… austria NA               NA        NA        
#>  3 spectacular_ameri… spectacular… france  NA               NA        NA        
#>  4 antimonarchy_cani… antimonarch… germany NA               NA        NA        
#>  5 nonphilosophical_… nonphilosop… germany NA               NA        NA        
#>  6 quasifaithful_amp… quasifaithf… germany NA               NA        NA        
#>  7 contrite_silkworm  contrite_si… germany NA               NA        NA        
#>  8 harmless_owlbutte… harmless_ow… germany NA               NA        NA        
#>  9 fascist_maiasaura  fascist_mai… germany NA               NA        NA        
#> 10 ironhearted_tarpan ironhearted… germany NA               NA        NA        
#> # ℹ 62 more rows
#> # ℹ 19 more variables: matched_activity_name <chr>,
#> #   matched_reference_product <chr>, co2e_lower <dbl>, co2e_upper <dbl>,
#> #   unit <chr>, multi_match <lgl>, matching_certainty <chr>,
#> #   matching_certainty_company_average <chr>, tilt_sector <chr>,
#> #   tilt_subsector <chr>, isic_4digit <chr>, isic_4digit_name <chr>,
#> #   company_city <chr>, postcode <chr>, address <chr>, main_activity <chr>, …

result |> unnest_company()
#> # A tibble: 72 × 13
#>    companies_id     company_name country emission_profile_share emission_profile
#>    <chr>            <chr>        <chr>                    <dbl> <lgl>           
#>  1 asteria_megalot… asteria_meg… austria                     NA NA              
#>  2 skarn_gallinule  skarn_galli… austria                     NA NA              
#>  3 spectacular_ame… spectacular… france                      NA NA              
#>  4 antimonarchy_ca… antimonarch… germany                     NA NA              
#>  5 nonphilosophica… nonphilosop… germany                     NA NA              
#>  6 quasifaithful_a… quasifaithf… germany                     NA NA              
#>  7 contrite_silkwo… contrite_si… germany                     NA NA              
#>  8 harmless_owlbut… harmless_ow… germany                     NA NA              
#>  9 fascist_maiasau… fascist_mai… germany                     NA NA              
#> 10 ironhearted_tar… ironhearted… germany                     NA NA              
#> # ℹ 62 more rows
#> # ℹ 8 more variables: benchmark <lgl>, co2e_lower <dbl>, co2e_upper <dbl>,
#> #   matching_certainty_company_average <chr>, company_city <chr>,
#> #   postcode <chr>, address <chr>, main_activity <chr>

inputs <- read_csv(toy_emissions_profile_upstream_products())
#> Warning: `toy_emissions_profile_upstream_products()` was deprecated in tiltToyData
#> 0.0.0.9009.
#> ℹ Please use `toy_emissions_profile_upstream_products_ecoinvent()` instead.
ecoinvent_inputs <- read_csv(toy_ecoinvent_inputs())

result <- profile_emissions_upstream(
  companies,
  inputs,
  europages_companies = europages_companies,
  ecoinvent_activities = ecoinvent_activities,
  ecoinvent_inputs = ecoinvent_inputs,
  ecoinvent_europages = ecoinvent_europages,
  isic = isic_name
)
#> Warning: Splitting `companies` into 4 chunks.

result |> unnest_product()
#> # A tibble: 72 × 28
#>    companies_id company_name country emission_upstream_pr…¹ benchmark ep_product
#>    <chr>        <chr>        <chr>   <chr>                  <lgl>     <chr>     
#>  1 asteria_meg… asteria_meg… austria NA                     NA        NA        
#>  2 skarn_galli… skarn_galli… austria NA                     NA        NA        
#>  3 spectacular… spectacular… france  NA                     NA        NA        
#>  4 antimonarch… antimonarch… germany NA                     NA        NA        
#>  5 nonphilosop… nonphilosop… germany NA                     NA        NA        
#>  6 quasifaithf… quasifaithf… germany NA                     NA        NA        
#>  7 contrite_si… contrite_si… germany NA                     NA        NA        
#>  8 harmless_ow… harmless_ow… germany NA                     NA        NA        
#>  9 fascist_mai… fascist_mai… germany NA                     NA        NA        
#> 10 ironhearted… ironhearted… germany NA                     NA        NA        
#> # ℹ 62 more rows
#> # ℹ abbreviated name: ¹​emission_upstream_profile
#> # ℹ 22 more variables: matched_activity_name <chr>,
#> #   matched_reference_product <chr>, co2e_lower <dbl>, co2e_upper <dbl>,
#> #   unit <chr>, multi_match <lgl>, matching_certainty <chr>,
#> #   matching_certainty_company_average <chr>, input_name <chr>,
#> #   input_unit <chr>, input_tilt_sector <chr>, input_tilt_subsector <chr>, …

result |> unnest_company()
#> # A tibble: 72 × 13
#>    companies_id         company_name company_city country emission_upstream_pr…¹
#>    <chr>                <chr>        <chr>        <chr>                    <dbl>
#>  1 asteria_megalotomus… asteria_meg… wilhelmsburg austria                     NA
#>  2 skarn_gallinule      skarn_galli… wiener neud… austria                     NA
#>  3 spectacular_america… spectacular… nice         france                      NA
#>  4 antimonarchy_canine  antimonarch… hamburg      germany                     NA
#>  5 nonphilosophical_ll… nonphilosop… solms        germany                     NA
#>  6 quasifaithful_amphi… quasifaithf… köln         germany                     NA
#>  7 contrite_silkworm    contrite_si… krefeld      germany                     NA
#>  8 harmless_owlbutterf… harmless_ow… edling       germany                     NA
#>  9 fascist_maiasaura    fascist_mai… bielefeld    germany                     NA
#> 10 ironhearted_tarpan   ironhearted… lindlar      germany                     NA
#> # ℹ 62 more rows
#> # ℹ abbreviated name: ¹​emission_upstream_profile_share
#> # ℹ 8 more variables: emission_upstream_profile <lgl>, benchmark <lgl>,
#> #   co2e_lower <dbl>, co2e_upper <dbl>,
#> #   matching_certainty_company_average <chr>, postcode <chr>, address <chr>,
#> #   main_activity <chr>